+++ tests/test_unit/test_core/test_template_manager.py
@@ -12,8 +12,9 @@ import pytest
 
 from mcp_template.core.template_manager import TemplateManager
 
+pytestmark = pytest.mark.unit
 
-@pytest.mark.unit
+
 class TestTemplateManager:
     """Unit tests for TemplateManager class."""
 
@@ -482,28 +483,6 @@ class TestTemplateManager:
                     config = self.template_manager.load_template_config("demo")
 
         assert config == mock_config
-
-
-@pytest.mark.integration
-class TestTemplateManagerIntegration:
-    """Integration tests for TemplateManager."""
-
-    def test_template_manager_with_real_templates(self):
-        """Test template manager with real template discovery."""
-        # This would test with actual template files
-        template_manager = TemplateManager(backend_type="mock")
-        templates = template_manager.list_templates()
-
-        # Should discover real templates in the system
-        assert isinstance(templates, dict)
-
-    def test_template_manager_with_mock_backend(self):
-        """Test template manager with mock backend."""
-        template_manager = TemplateManager(backend_type="mock")
-
-        # Should be able to list templates without errors
-        templates = template_manager.list_templates(include_deployed_status=True)
-        assert isinstance(templates, dict)
 
 
 def mock_open_read_json(json_data):
